change models to use enums; add byte conversion in context for db storage

+++ DemoCentral/Models/MatchHistoryModel.cs
@@ -20,7 +20,7 @@ namespace DemoCentral.Models
                 {
                     MatchId = match.MatchId,
                     MatchDate = match.MatchDate,
-                    Success = match.DemoAnalyzerStatus == (byte) DemoAnalyzerStatus.Finished
+                    Success = match.DemoAnalyzerStatus == DemoAnalyzerStatus.Finished
                 });
             }
